+++ src/cli/test.rs
@@ -126,6 +126,13 @@ fn test_one(test: &TestCase) {
     let temp = tempfile::tempdir().expect("Failed creating a temp directory");
     std::env::set_current_dir(temp.path()).expect("Failed chainging current directory");
 
+    tarantool::exec(
+        "require 'compat' {
+            c_func_iproto_multireturn = 'new',
+        }",
+    )
+    .unwrap();
+
     let cfg = tarantool::Cfg {
         listen: Some("127.0.0.1:0".into()),
         read_only: false,